+++ test/unit/driver_test.rb
@@ -230,6 +230,58 @@ class AppiumLibCoreTest
       assert_equal '/wd/hub/', uri.path
     end
 
+    def test_default_timeout_for_http_client_with_direct_no_supported_client
+      def android_mock_create_session_w3c_direct_default_client(core)
+        response = {
+          value: {
+            sessionId: '1234567890',
+            capabilities: {
+              platformName: :android,
+              automationName: ENV['AUTOMATION_NAME_DROID'] || 'uiautomator2',
+              app: 'test/functional/app/api.apk.zip',
+              platformVersion: '7.1.1',
+              deviceName: 'Android Emulator',
+              appPackage: 'io.appium.android.apis',
+              appActivity: 'io.appium.android.apis.ApiDemos',
+              someCapability: 'some_capability',
+              unicodeKeyboard: true,
+              resetKeyboard: true,
+              directConnectProtocol: 'http',
+              directConnectHost: 'localhost',
+              directConnectPort: '8888',
+              directConnectPath: '/wd/hub'
+            }
+          }
+        }.to_json
+
+        stub_request(:post, 'http://127.0.0.1:4723/wd/hub/session')
+          .to_return(headers: HEADER, status: 200, body: response)
+
+        stub_request(:post, 'http://127.0.0.1:4723/wd/hub/session/1234567890/timeouts')
+          .with(body: { implicit: 30_000 }.to_json)
+          .to_return(headers: HEADER, status: 200, body: { value: nil }.to_json)
+
+        driver = core.start_driver http_client_ops: { http_client: Selenium::WebDriver::Remote::Http::Default.new }
+
+        assert_requested(:post, 'http://127.0.0.1:4723/wd/hub/session', times: 1)
+        assert_requested(:post, 'http://127.0.0.1:4723/wd/hub/session/1234567890/timeouts',
+                         body: { implicit: 30_000 }.to_json, times: 1)
+        driver
+      end
+
+      core = ::Appium::Core.for(Caps.android_direct)
+      driver = android_mock_create_session_w3c_direct_default_client(core)
+
+      assert_nil driver.send(:bridge).http.open_timeout
+      assert_nil driver.send(:bridge).http.read_timeout
+      uri = driver.send(:bridge).http.send(:server_url)
+      assert core.direct_connect
+      assert_equal 'http', uri.scheme
+      assert_equal '127.0.0.1', uri.host
+      assert_equal 4723, uri.port
+      assert_equal '/wd/hub/', uri.path
+    end
+
     def test_default_timeout_for_http_client_with_enable_idempotency_header_false
       def _android_mock_create_session_w3c(core)
         response = {